Work in progress for implementation of frontend in HTML and CSS
Took original design and tried to implement in HTML and CSS. Can be used as a basis for the final frontend we develop. So far only the right side bar is implemented, will update and implement entire design.

Features are implemented more so to test visual design, functionality with API and individual style elements are to be added later. As a result some elements exist only as placeholders.

All CSS is done inline or internally. If an external CSS file is preferred I will add it as a task on the issue page for the text system frontend.

Measurements are done using viewportwidth (vw) and viewportheight (vh) instead of percentages as the percentages did not work when I tested it.